requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
gradio>=3.40.0
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.5.0
datasets>=2.12.0
nltk>=3.8
sacrebleu>=2.3.0
pycocoevalcap>=1.2
googletrans>=4.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
requests>=2.28.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
accelerate>=0.20.0
